connect: connect CA Roots in secondary datacenters should use a SigningKeyID derived from their local intermediate
This fixes an issue where leaf certificates issued in secondary
datacenters would be reissued very frequently (every ~20 seconds)
because the logic meant to detect root rotation was errantly triggering
because a hash of the ultimate root (in the primary) was being compared
against a hash of the local intermediate root (in the secondary) and
always failing.

This does not have support yet to correct previously stored CARoot
objects.